Очередные правки и доделки... Пересоздана таблица payment_options (добавлены колонки для оплаты заказа и комиссия оплаты) Таблица orders (теперь цепляется к payment_options) Таблица order_items (исправлены типы полей Цены и Процент комиссии) Удалены лишние миграции Создана миграция для истории счета продавца Создана миграция для холдов Создана таблица входящих/исходящих оплат Переписан расчет комиссий пополнения/вывода Новый класс приема платежей Qiwi (от Дена). Старые удалены При оформлении заказа способы оплат берутся из таблицы payment_options

+++ database/migrations/2020_07_01_191949_create_payment_schedules_table.php
@@ -15,7 +15,7 @@ class CreatePaymentSchedulesTable extends Migration
     {
         Schema::create('payment_schedules', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('percent');
+            $table->unsignedTinyInteger('percent');
             $table->integer('quantity_pay_every_month');
             $table->integer('quantity_pay_each_quarter');
             $table->integer('quantity_pay_every_six_months');
